pyproject.toml
[tool.poetry]
name = "star-types"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8.6"
pandas = "^1.2.2"
sklearn = "^0.0"
matplotlib = "^3.3.4"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
